feat: update/delete product

+++ src/main/java/com/parmar/explore/controller/ProductController.java
@@ -5,7 +5,6 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -13,7 +12,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
@@ -53,7 +51,7 @@ public class ProductController {
 
     }
 
-    @PostMapping("/products")
+    @PostMapping("/product")
     public ResponseEntity<?> addProduct(@RequestPart Product product, @RequestPart MultipartFile imageFile) {
         try {
             Product prod = service.addProduct(product, imageFile);
@@ -68,17 +66,32 @@ public class ProductController {
         Product product = service.getProductById(prodId);
         byte[] imageFile = product.getImageDate();
 
-        return ResponseEntity.ok().contentType(MediaType.valueOf(product.getImageType())).body(imageFile);
+        return ResponseEntity.ok()
+                // .contentType(MediaType.valueOf(product.getImageType()))
+                .body(imageFile);
     }
 
-    @PutMapping("products")
-    public List<Product> updateProduct(@RequestBody Product product) {
-        service.updateProduct(product);
-        return service.getAllProducts();
+    @PutMapping("product/{id}")
+    public ResponseEntity<String> updateProduct(@PathVariable int id, @RequestPart Product product, @RequestPart MultipartFile imageFile) throws IOException {
+        Product product1 = service.updateProduct(id, product, imageFile);
+
+        if (product1 != null) {
+            return new ResponseEntity<>("Product Updated Successfully", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Product failed to update", HttpStatus.BAD_REQUEST);
+        }
     }
 
     @DeleteMapping("product/{prodId}")
-    public void deleteProduct(@PathVariable int prodId) {
-        service.deleteProduct(prodId);
+    public ResponseEntity<String> deleteProduct(@PathVariable int prodId) {
+        Product product = service.getProductById(prodId);
+
+        if (product != null) {
+            service.deleteProduct(prodId);
+            return new ResponseEntity<>("Product Deleted Successfully", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Product failed to delete", HttpStatus.BAD_REQUEST);
+        }
+
     }
 }
